Module lambek::refl

Implementation for the type quality constraint, a.k.a. reflexivity.

This module introduces the Refl trait with an associated type Refl, which is implemented for all types with the Refl::Refl being the same type that implements Refl.

trait Refl {
    type Refl;
}

impl <T> Refl for T {
    type Refl = T;
}

This means that if we have T1: Refl<Refl=T2>, that means that T1 and T2 are in fact the same type and can be used interchangeably.

The Refl trait by itself do not have much use, as Rust currently have amnesia on the concrete type of an associated type. For example, the following code would fail to compile, even though conceptually it should succeed:

use lambek::refl::Refl;

fn reflect_self<T1, T2>(x: T1) -> T2
where
    T1: Refl<Refl=T2>,
{
    x
}

We have already know that T1 implements Refl and T1::Refl must be also T1, so T2 must be equal to T1 and so we should be able to return x as T2 instead of T1. But the Rust compiler currently have amnesia and forgot the concrete type of T1::Refl. So it cannot conclude that T1 and T2 are equal.

Fortunately, there is in fact a way for us to help Rust recover from amnesia, and remember back the concrete type for T1::Refl. We do this by defining a trait ReflSelf that extends Refl, with a generic implementation of how to reflect back itself:

trait ReflSelf: Refl {
    fn refl_self(self) -> Self::Refl;
}

impl <T> ReflSelf for T {
    fn refl_self(self) -> Self::Refl {
        self
    }
}

fn reflect_self<T1, T2>(x: T1) -> T2
where
    T1: Refl<Refl=T2>,
{
    T1::refl_self(x)
}

The trait ReflSelf implements Refl and also provide a method refl_self that accepts a self and returns Self::Refl. Since we know Self::Refl is in fact the same as Self, the implementation is trivial with us simply returning self.

Notice that in the generic implementation of ReflSelf, we state that the trait is implemented for all types. This is allowed because in the trait implementation scope, Rust do look up the trait implementation for Refl, and find out that Refl is implemented for all types. In other words, at this stage, Rust have not have amnesia and still remember that the concrete type for Refl::Refl is in fact Self. Using that fact, we easily implement refl_self to return self without Rust complaining.

Using ReflSelf, we help put a sticky note in Rust’s mind, and we can use that to remind Rust again what the type of Refl::Refl is when we need it. We now try to implement reflect_self again, which now calls ReflSelf::refl_self to convert T1 to T1::Refl. But since we also specify the constraint T1: Refl<Refl=T2>, that means Rust actually thinks that the type T1::Refl returned by ReflSelf::refl_self must also be T2, and therefore the code compiles successfully.

Using the same technique, we can also construct new traits that extend Refl to perform any reflexivity conversion that we need. For example, let’s say we have a Vec<T1> we know that T1: Refl<Refl=T2>. Following that we should be able to convert our Vec<T1> into a Vec<T2>. So we can define a ReflVec trait like follows:

use lambek::refl::Refl;

pub trait ReflVec: Refl {
    fn refl_vec(left: Vec<Self>) -> Vec<Self::Refl>;
}

impl <X> ReflVec for X {
    fn refl_vec(left: Vec<Self>) -> Vec<Self::Refl> {
        left
    }
}

pub fn reflect_vec<T1, T2>(xs: Vec<T1>) -> Vec<T2>
where
    T1: Refl<Refl=T2>,
{
    T1::refl_vec(xs)
}

We can also use the technique to cast anything, including references and boxed values, so that the conversion can be done at zero cost. lambek also provides the ReflApp trait, so that users can easily reflect any higher kinded types using the proxy types that implement TypeApp.

Traits

HasRefl

HasReflUnrestricted

Refl

ReflApp

Reflect higher kinded types when there is a type equality for the argument types.

ReflSelf

Extend Refl to provide reflection on base values.

Functions

app_congruence

has_refl

option_congruence

option_ref_congruence

refl_symmetric

refl_transitive

reflect_applied

reflect_boxed

reflect_right_ref

reflect_value

reflect_value_mut

reflect_value_ref

reflect_value_right

vec_congruence