Trait lambek::refl::ReflApp

pub trait ReflApp: Refl {
    fn refl_app<'a, F>(
        applied: <F as TypeApp<'a, Self>>::Applied
    ) -> <F as TypeApp<'a, Self::Refl>>::Applied
    where
        Self: 'a,
        F: TypeApp<'a, Self>,
        F: TypeApp<'a, Self::Refl>,
        <F as TypeApp<'a, Self>>::Applied: Sized,
        <F as TypeApp<'a, Self::Refl>>::Applied: Sized;
}

Reflect higher kinded types when there is a type equality for the argument types.

We make use of the type_app module which define high kinded types to help us reflect type equalities within other types.

For example, since Applied<OptionF, T1> is the same as Option<T1>, we can use OptionF::refl_app to reflect a Option<T1> into Option<T2> if T1: Refl<Refl=T2>:

use lambek::refl::{Refl, ReflApp};
use lambek::type_app::OptionF;

pub fn reflect_option<T1, T2>(xs: Option<T1>) -> Option<T2>
where
    T1: Refl<Refl = T2>,
{
    T1::refl_app::<OptionF>(xs)
}

For reflecting more complex types such as &Option<T1>, we can also use OptionF together with the HKT combinator Compose and the Borrow type. Together, the type Applied<Compose<Borrow, OptionF>, T1> is the same as &Option<T1>. Hence we can use ReflApp to reflect that with Compose<Borrow, OptionF> being the HKT reflector:

use lambek::refl::{Refl, ReflApp};
use lambek::type_app::{Compose, OptionF};
use lambek::reference::Borrow;

pub fn reflect_option_borrow<T1, T2>(xs: &Option<T1>) -> &Option<T2>
where
    T1: Refl<Refl = T2>,
{
    T1::refl_app::<Compose<Borrow, OptionF>>(xs)
}

Required methods

fn refl_app<'a, F>(
applied: <F as TypeApp<'a, Self>>::Applied
) -> <F as TypeApp<'a, Self::Refl>>::Applied where
Self: 'a,
F: TypeApp<'a, Self>,
F: TypeApp<'a, Self::Refl>,
<F as TypeApp<'a, Self>>::Applied: Sized,
<F as TypeApp<'a, Self::Refl>>::Applied: Sized,