Trait lambek::type_app::base::TypeApp

pub trait TypeApp<'a, X: 'a + ?Sized>: TypeCon + 'a {
    type Applied: 'a + ?Sized;
}

A type F: TypeApp<X> have the associated type Applied as the result of applying a type F of kind Type -> Type to X.

More specifically, TypeApp is also parameterized by a lifetime 'a to support application of types with limited lifetime. Unlike other functional languages, the higher kinded type application F X have to consider the case that both F and X may have different lifetimes. To deal with that, we require that a type X can only be applied to a type F if they both share a common lifetime 'a. The result of the type application must also have the lifetime 'a.

In practice, we typically define F to have 'static lifetime, i.e. they do not contain references. On the other hand the type argument X may contain references. For example, the result of VecF: TypeApp<'a, &'aX> would be Vec<&'a X>. A typical implementation of TypeApp would something like follows:

enum FooF {}
struct Foo<X>(X);
impl TypeCon for FooF {}
impl<'a, X: 'a> TypeApp<'a, X> for FooF
{
  type Applied = Foo<X>;
}

A type constructor F may also choose to implement TypeApp for unsized type arguments X, e.g. dyn trait objects. For example, we could define a type BarF to make the result of applying a type X into dyn Bar<X>:

enum BarF {}
trait Bar<X>
{
  fn bar(
    &self,
    x: X,
  );
}
impl TypeCon for BarF {}
impl<'a, X: 'a> TypeApp<'a, X> for BarF
{
  type Applied = dyn Bar<X> + 'a;
}

Associated Types

type Applied: 'a + ?Sized